chore(marketing): /solve deploy fixes + vercel.json + rate-limit module (AGE-104)
Tightens up the AGE-104 /solve survey deploy to Vercel.

## Changes

- `marketing/vercel.json` — pin framework=nextjs and explicit build/dev
  commands. Vercel auto-detected the project as static after a fresh
  `vercel project add agentdash`, missing the Next.js framework hint.

- `marketing/.npmrc` — `ignore-workspace=true` so Vercel's pnpm install
  doesn't try to resolve the parent monorepo workspace (marketing/ has
  no `@agentdash/*` deps; treating it as standalone here is the simplest
  install path).

- `marketing/src/lib/rate-limit.ts` — extract the in-memory IP rate
  limiter out of the route handler. Vercel rejects non-Route exports
  from `route.ts` files at build time
  (`__resetRateLimit is not a valid Route export field`). The new
  module exposes `rateLimit()` for runtime + `_resetRateLimitForTests()`
  for the unit test.

- `marketing/src/app/api/solve-submit/route.ts` — import `rateLimit`
  from the new module; drop the inline copy + the test-only export.

- `marketing/src/app/api/solve-submit/__tests__/route.test.ts` — import
  `_resetRateLimitForTests` from the new module.

- `marketing/src/lib/solve-store.ts` — comment-only clarification on
  the @vercel/blob `access: "public"` parameter (the SDK requires
  `"public"` literal even when the store is private — store-side ACL
  governs effective access).

- `marketing/.gitignore` — add `.vercel/` (added by `vercel link`).

## Deployment state

The /solve survey is live at https://agentdash-steel.vercel.app/solve
with these confirmed working:

- Vercel project `thetangstrs-projects/agentdash`
- Vercel Blob store `agentdash-solve` (public, unguessable UUIDs as
  paths) linked to all 3 environments
- Resend env vars (`RESEND_API_KEY`, `RESEND_FROM_EMAIL`,
  `RESEND_OPS_EMAIL`) set on Production + Development
- Smoke submission verified durable in Blob:
  `submissions/2026-05-01/231c889a-...json`
