+++ marketing/.gitignore
@@ -3,3 +3,4 @@ node_modules/
 out/
 *.log
 .env*.local
+.vercel
