+++ marketing/.npmrc
@@ -1,0 +1,5 @@
+# Vercel build: marketing/ is technically inside a pnpm workspace, but it has
+# zero workspace deps. Tell pnpm to ignore the workspace context when
+# installing here so Vercel's deploy can succeed without needing to upload
+# the entire monorepo.
+ignore-workspace=true
