+++ marketing/vercel.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "https://openapi.vercel.sh/vercel.json",
+  "framework": "nextjs",
+  "buildCommand": "next build",
+  "devCommand": "next dev -p 3200",
+  "ignoreCommand": "git diff --quiet HEAD^ HEAD -- ."
+}
